+++ attoparsec-iso8601/changelog.md
@@ -1,5 +1,9 @@
 For the latest version of this document, please see [https://github.com/bos/aeson/blob/master/attoparsec-iso8601/changelog.md](https://github.com/bos/aeson/blob/master/attoparsec-iso8601/changelog.md).
 
+### 1.0.2.0
+
+* Add `month :: Parser Month` and `quarter :: Parser Quarter`
+
 ### 1.0.1.0
 
 * Fixes handling of `UTCTime` wrt. leap seconds , thanks to Adam Schønemann